Fix: LvmDriver
Commit `fa5d44e939839bd119273a817a51bb0cf2c066bc` introduced a bug where
deleteVolume looks for the volume which should be deleted. In case the
volume could not be created in the first place, that lookup was
configured to throw an exception because it didnt find the volume.

+++ src/com/linbit/linstor/storage/LvmDriver.java
@@ -109,9 +109,12 @@ public class LvmDriver extends AbsStorageDriver
     @Override
     public void deleteVolume(String identifier, boolean isEncrypted) throws StorageException
     {
-        final VolumeInfo info = getVolumeInfo(identifier);
-        callDmStat("dmstats", "delete", info.getPath(), "--allregions");
-        super.deleteVolume(identifier, isEncrypted);
+        final VolumeInfo info = getVolumeInfo(identifier, false);
+        if (info != null)
+        {
+            callDmStat("dmstats", "delete", info.getPath(), "--allregions");
+            super.deleteVolume(identifier, isEncrypted);
+        }
     }
 
     private void callDmStat(String... command) throws StorageException
